// flask-q4-recommendations-demo/new-q4-app/src/API.ts
export enum AnnotationTaskStatus {
  TODO = "TODO",
  FAILED = "FAILED",
  DONE = "DONE",
}


export type NerAnnotationSampleInput = {
  text: string,
  annotations?: Array< NerAnnotationInput | null > | null,
};

export type NerAnnotationInput = {
  text: string,
  tag?: AnnotationTagInput | null,
  start: number,
  end: number,
};

export type AnnotationTagInput = {
  id: string,
  name: string,
  description: string,
};
// ...
export type AnnotationTaskInstruction = {
  __typename: "AnnotationTaskInstruction",
  text: string,
  full_instructions_url?: string | null,
};

export enum AnnotationTaskType {
  NER = "NER",
  CLASSIFICATION = "CLASSIFICATION",
  ALLERGIES = "ALLERGIES",
}


export type AnnotationTag = {
  __typename: "AnnotationTag",
  id: string,
  name: string,
  description: string,
};

export type ClassificationAnnotationSample = {
  __typename: "ClassificationAnnotationSample",
  text: string,
  annotations:  Array<ClassificationAnnotation | null >,
};

export type ClassificationAnnotation = {
  __typename: "ClassificationAnnotation",
  text: string,
  tag?: AnnotationTag | null,
};

export type AnnotationTaskIssue = {
  __typename: "AnnotationTaskIssue",
  id: string,
  type: AnnotationTaskIssueType,
  reason?: AnnotationTaskIssueReason | null,
  notes: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueType = {
  __typename: "AnnotationTaskIssueType",
  id: string,
  type: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueReason = {
  __typename: "AnnotationTaskIssueReason",
  id: string,
  reason: string,
  createdAt: string,
  updatedAt: string,
};

export type AllergyAnnotationTask = {
  __typename: "AllergyAnnotationTask",
  id: string,
  project_id: string,
  instruction: AnnotationTaskInstruction,
  status: AnnotationTaskStatus,
  type: AnnotationTaskType,
  metadata?: string | null,
  annotator_id: string,
  tags:  Array<AnnotationTag >,
  sample: AllergyAnnotationSample,
  issue?: AnnotationTaskIssue | null,
  createdAt: string,
  updatedAt: string,
};

export type AllergyAnnotationSample = {
  __typename: "AllergyAnnotationSample",
  text: string,
  annotations?:  Array<AnnotationTagWithEvidence | null > | null,
};

export type AnnotationTagWithEvidence = {
  __typename: "AnnotationTagWithEvidence",
  id: string,
  tag?: AnnotationTag | null,
  evidence?: ModelEvidenceConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelEvidenceConnection = {
  __typename: "ModelEvidenceConnection",
  items?:  Array<Evidence | null > | null,
  nextToken?: string | null,
};
// ...
export type AnnotationTaskInstructionInput = {
  text: string,
  full_instructions_url?: string | null,
};
// ...
export type CreateAllergyAnnotationTaskInput = {
  id?: string | null,
  project_id: string,
  instruction: AnnotationTaskInstructionInput,
  status: AnnotationTaskStatus,
  type: AnnotationTaskType,
  metadata?: string | null,
  annotator_id: string,
  tags: Array< AnnotationTagInput >,
  sample: AllergyAnnotationSampleInput,
  allergyAnnotationTaskIssueId?: string | null,
};

export type AllergyAnnotationSampleInput = {
  text: string,
};